# swagger.yaml
openapi: 3.0.0
info:
  title: Submissions API
  contact: {}
  version: '1.0'
servers:
  - url: https://acolyte-submissions-api.herokuapp.com/api/v1
    variables: {}
paths:
  /auth/delegate:
    post:
      tags:
        - Auth
      summary: Add Reader
      operationId: Delegate
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DelegateRequest'
            example:
              name: bob
              email: bob@gmail.com
              password: secret
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
...
  /submissions:
    post:
      tags:
        - Submissions
      summary: Create Submitted
      operationId: CreateSubmitted
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubmissionRequest'
            example:
              company: google
              position: front-end developer
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
        - Submissions
      summary: Get All Submissions
      operationId: GetAllSubmissions
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
components:
  schemas:
    DelegateRequest:
      title: DelegateRequest
      required:
        - name
        - email
        - password
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      example:
        name: bob
        email: bob@gmail.com
        password: secret
...
    CreateSubmissionRequest:
      title: CreateSubmissionRequest
      required:
        - company
        - position
      type: object
      properties:
        company:
          type: string
        position:
          type: string
      example:
        company: google
        position: front-end developer
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
  - httpBearer: []
...